Force conversion to smaller int type, because we know it must fit.

+++ include/osmium/io/detail/pbf_decoder.hpp
@@ -225,7 +225,7 @@ namespace osmium {
                 }
 
                 int32_t convert_pbf_coordinate(int64_t c) const {
-                    return (c * m_granularity + m_lon_offset) / resolution_convert;
+                    return int32_t((c * m_granularity + m_lon_offset) / resolution_convert);
                 }
 
                 void decode_node(const ptr_len_type& data) {
